Добавлена работа с отзывами.
- Добавлено взаимодействие с отзывами
- TODO: покрыть тестами, документация

+++ DomovoiBackend.Application/Requests/Reviews/AddReviewRequest.cs
@@ -4,7 +4,7 @@ public class AddReviewRequest
 {
     public Guid DestinationId { get; set; }
     public Guid AuthorId { get; set; }
-    public byte Rate { get; set; }
+    public int Rate { get; set; }
     public string Header { get; set; } = null!;
     public string Text { get; set; } = null!;
 }
